requirements.txt
faiss_gpu==1.7.2
numpy==1.24.2
pandas==1.5.3
sentence_transformers==2.2.2
torch==1.13.0
tqdm==4.65.0
transformers==4.43.0
